+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - trigger support
 - webos video playback hotfix with sleep(videoDuration) and Promise.race
 - add priority behaviour support (excl and priorityClass tags)
+- add conditional playback support
 
 ### Fixed
 - fixed smil parsing issue on Rpi